extract a PlanBenefits module
